+++ CHANGELOG.md
@@ -1,2 +1,5 @@
+## 0.0.2
+Added possibility to override the Region from which to list certificates.
+This is useful whe getting a certificate for CloudFront which have to be issued in us-east-1
 ## 0.0.1
 Initial release